+++ src/schemas/index.ts
@@ -1,6 +1,6 @@
 export { personSchema, type PersonOptions } from './person';
 export { organizationSchema, type OrganizationOptions } from './organization';
-export { websiteSchema, type WebSiteOptions } from './webSite';
+export { webSiteSchema, type WebSiteOptions } from './webSite';
 export { webPageSchema, aboutPageSchema, contactPageSchema, type WebPageOptions } from './webPage';
 export { webApiSchema, type WebApiOptions } from './webApi';
 export { softwareAppSchema, type SoftwareAppOptions } from './softwareApp';